header: make a component for the profile drop down menu
