+++ src/components/Header.tsx
@@ -1,14 +1,12 @@
 import Button from "./Button";
 import { IoMdNotificationsOutline } from "react-icons/io";
 import { RiMenu4Fill } from "react-icons/ri";
-import Logo from "../../public/logo/EventSync.svg";
 import Image from "next/image";
-import { FaRegCircleUser } from "react-icons/fa6";
-import { MdOutlineLogout } from "react-icons/md";
 import { useState } from "react";
 import { useSession } from "next-auth/react";
 import { api } from "~/utils/api";
 import { useRouter } from "next/router";
+import HeaderProfileIconMenu from "./HeaderProfileIconMenu";
 
 const Header = () => {
   const router = useRouter();
@@ -24,14 +22,6 @@ const Header = () => {
     /* TODO redirect to a page  */
     console.log("manage event button onClick", route);
     await router.push(route);
-  };
-  const onClickIconMenu = () => {
-    setOpenIconMenu(!openIconMenu);
-    console.log("onclick menu");
-  };
-
-  const logout = () => {
-    console.log("logout");
   };
 
   return (
@@ -62,36 +52,7 @@ const Header = () => {
             onClick={() => goTo("/notification")}
           />
           {sessionData ? (
-            <div className="relative">
-              <div
-                className="h-10 w-10 cursor-pointer rounded-full border bg-black text-white hover:shadow-lg"
-                onMouseEnter={onClickIconMenu}
-              >
-                <p className="mt-2 text-center">W</p>
-              </div>
-              {openIconMenu && (
-                <div
-                  className="absolute right-0 top-14 min-w-44 p-5 text-gray-700 shadow-lg"
-                  onMouseLeave={onClickIconMenu}
-                >
-                  <div
-                    className="hover:text-blue flex cursor-pointer items-center justify-between hover:text-cyan-600"
-                    onClick={() => goTo("/profile")}
-                  >
-                    <FaRegCircleUser />
-                    <p>My Profile</p>
-                  </div>
-                  <hr className="my-3"></hr>
-                  <div
-                    className="flex cursor-pointer items-center justify-between hover:text-cyan-600"
-                    onClick={logout}
-                  >
-                    <MdOutlineLogout />
-                    <p>Sign out</p>
-                  </div>
-                </div>
-              )}
-            </div>
+            <HeaderProfileIconMenu/>
           ) : (
             <div className="cursor-pointer" onClick={() => goTo("/login")}>
               Sign In
